requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
python-jose==3.3.0
cryptography==35.0.0
django-cors-headers==3.10.0
mysqlclient==2.0.3